pyproject.toml
[project]
name = "agent-observability"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.2",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-exporter-prometheus>=0.58b0",
    "opentelemetry-instrumentation-logging>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "prometheus-client>=0.23.1",
    "uvicorn>=0.36.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
api = "agent_observability:main.run_api"
